Implement working app functionality
