+++ Movie List/Base.lproj/Main.storyboard
@@ -55,7 +55,7 @@
         <!--Movies-->
         <scene sceneID="1he-q7-G3K">
             <objects>
-                <viewController id="BfI-d4-uht" userLabel="Movies" customClass="ViewController" customModule="Movie_List" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="BfI-d4-uht" userLabel="Movies" customClass="TableViewController" customModule="Movie_List" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="IJN-ET-3Bh">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -86,7 +86,7 @@
                     </view>
                     <tabBarItem key="tabBarItem" title="Movies" image="Clapboard" id="eg9-qe-slp"/>
                     <connections>
-                        <outlet property="tableView" destination="wXN-5K-Y31" id="e7t-cF-0vO"/>
+                        <outlet property="tableView" destination="wXN-5K-Y31" id="NYI-jE-KBB"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="IZr-NY-I8a" userLabel="First Responder" sceneMemberID="firstResponder"/>
