+++ Movie List/ViewController.swift
@@ -1,46 +1,21 @@
 import UIKit
 
-class ViewController: UIViewController, UITableViewDataSource, UITableViewDelegate {
+class ViewController: UIViewController {
     
     @IBOutlet weak var textField: UITextField!
-    
-    @IBOutlet weak var tableView: UITableView!
     
     @IBAction func addMovie(_ sender: Any) {
         guard let text = textField.text, !text.isEmpty else { return }
         Model.shared.addMovie(text)
         textField.text = nil
-        tableView.insertRows(at: [IndexPath(row: Model.shared.movieCount() - 1, section: 0)], with: .top)
+        // TableViewController.tableView.insertRows(at: [IndexPath(row: Model.shared.movieCount() - 1, section: 0)], with: .top)
     }
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        tableView.dataSource = self
-        tableView.delegate = self
-        tableView.reloadData()
+        // tableView.dataSource = self
+        // tableView.delegate = self
+        // tableView.reloadData()
     }
     
-    let reuseIdentifier = "cell"
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return Model.shared.movieCount()
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: reuseIdentifier, for: indexPath)
-        cell.textLabel?.text = Model.shared.movie(at: indexPath.row)
-        return cell
-    }
-    
-    func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
-        // Enable "magic" swipe-to-delete functionality.
-        guard editingStyle == .delete else { return }
-        Model.shared.removeMovie(at: indexPath.row)
-        tableView.deleteRows(at: [indexPath], with: .fade)
-    }
-    
-    func tableView(_ tableView: UITableView, moveRowAt sourceIndexPath: IndexPath, to destinationIndexPath: IndexPath) {
-        Model.shared.moveMovie(from: sourceIndexPath.row, to: destinationIndexPath.row)
-        tableView.moveRow(at: sourceIndexPath, to: destinationIndexPath)
-    }
-    
+   
 }
